experiments/requirements.txt
annotated-types==0.7.0
humanfriendly==10.0
dateparser==1.2.1
autopep8==2.3.2
black==25.1.0
isort==6.0.1
kubernetes==32.0.1
pydantic==2.11.4
pydantic_core==2.33.2
pylint==3.3.7
pytest==8.3.5
PyYAML==6.0.2
regex==2024.11.6
ruamel.yaml==0.17.21
yamlpath==3.8.2
